requirements.txt
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
future==0.18.2
idna==2.10
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-twitter==3.5
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
